Fix plain text row limit typing

+++ matrix-jupyter/src/main/groovy/se/alipsa/matrix/jupyter/render/CoreRenderer.groovy
@@ -45,7 +45,7 @@ class CoreRenderer extends AbstractRenderer {
   private static String plainText(Matrix matrix, RenderOptions options, List<String> notices) {
     Matrix rendered = options.maxColumns != null && matrix.columnCount() > options.maxColumns ?
         matrix.selectColumns(matrix.columnNames().take(options.maxColumns)) : matrix
-    int rows = (options.maxRows != null ? options.maxRows : rendered.rowCount()).min(rendered.rowCount())
+    int rows = (int) (options.maxRows != null ? options.maxRows : rendered.rowCount()).min(rendered.rowCount())
     String title = rendered.toString() + '\n'
     String suffix = notices ? "\n${notices.join(', ')}" : ''
     String body = options.fromHead ? rendered.head(rows) : rendered.tail(rows)
